customers/requirements.txt
Django
djangorestframework
djoser
djangorestframework_simplejwt
mysqlclient
django-cors-headers
pika
python-dotenv
pytest-django
pytest-factoryboy
Faker
pytest-cov
requests
pyjwt